+++ fcli-core/fcli-app/src/main/java/com/fortify/cli/app/runner/util/FortifyCLIDynamicInitializer.java
@@ -15,6 +15,8 @@ package com.fortify.cli.app.runner.util;
 import java.io.File;
 import java.io.PrintWriter;
 import java.io.StringWriter;
+import java.util.HashMap;
+import java.util.Map;
 import java.util.function.Consumer;
 import java.util.stream.Stream;
 
@@ -23,14 +25,19 @@ import org.slf4j.LoggerFactory;
 import com.fortify.cli.common.cli.cmd.AbstractRunnableCommand;
 import com.fortify.cli.common.cli.cmd.AbstractRunnableCommand.GenericOptionsArgGroup;
 import com.fortify.cli.common.cli.cmd.AbstractRunnableCommand.LogLevel;
+import com.fortify.cli.common.log.LogMaskHelper;
+import com.fortify.cli.common.log.LogMessageType;
+import com.fortify.cli.common.log.LogMessageTypeConverter;
+import com.fortify.cli.common.log.LogSensitivityLevel;
+import com.fortify.cli.common.log.MaskingLogMessageConverter;
 import com.fortify.cli.common.util.DebugHelper;
 
 import ch.qos.logback.classic.Logger;
 import ch.qos.logback.classic.LoggerContext;
-import ch.qos.logback.classic.layout.TTLLLayout;
+import ch.qos.logback.classic.encoder.PatternLayoutEncoder;
 import ch.qos.logback.classic.spi.ILoggingEvent;
+import ch.qos.logback.core.CoreConstants;
 import ch.qos.logback.core.FileAppender;
-import ch.qos.logback.core.encoder.LayoutWrappingEncoder;
 import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -70,6 +77,8 @@ public final class FortifyCLIDynamicInitializer {
     }
         
     public void initializeLogging(GenericOptionsArgGroup genericOptions) {
+        LogMaskHelper.INSTANCE.setLogMaskLevel(genericOptions.getLogMaskLevel());
+        registerDefaultLogMaskPatterns();
         boolean isDebugEnabled = genericOptions.isDebug();
         File logFile = genericOptions.getLogFile();
         LogLevel logLevel = genericOptions.getLogLevel();
@@ -81,11 +90,30 @@ public final class FortifyCLIDynamicInitializer {
         if ( logLevel!=LogLevel.NONE && (logFile!=null || logLevel!=null) ) {
             // Configure logging if logLevel is not set to NONE, and logFile and/or logLevel 
             // have been specified
-            LoggerContext loggerContext = (LoggerContext) LoggerFactory.getILoggerFactory();
+            LoggerContext loggerContext = initializeLoggerContext();
             Logger rootLogger = loggerContext.getLogger(org.slf4j.Logger.ROOT_LOGGER_NAME);
             configureLogFile(rootLogger, logFile==null ? "fcli.log" : logFile.getAbsolutePath());
             configureLogLevel(rootLogger, logLevel==null ? LogLevel.INFO : logLevel);
         }
+    }
+
+    private void registerDefaultLogMaskPatterns() {
+        LogMaskHelper.INSTANCE
+            .registerPattern(LogSensitivityLevel.high, "Authorization: (?:[a-zA-Z]+ )?(.*?)(?:\\Q[\\r]\\E|\\Q[\\n]\\E)*\\\"?$", "<REDACTED>", LogMessageType.HTTP_OUT)
+            .registerPattern(LogSensitivityLevel.high, "(?:\\\"token\\\"|\\\"access_token\\\"):\\s*\\\"(.*?)\\\"", "<REDACTED TOKEN (RESPONSE)>", LogMessageType.HTTP_IN);
+    }
+
+    @SuppressWarnings("unchecked")
+    private LoggerContext initializeLoggerContext() {
+        var lc = (LoggerContext) LoggerFactory.getILoggerFactory();
+        var ruleRegistry = (Map<String, String>)lc.getObject(CoreConstants.PATTERN_RULE_REGISTRY);
+        if (ruleRegistry == null) {
+            ruleRegistry = new HashMap<String, String>();
+        }
+        lc.putObject(CoreConstants.PATTERN_RULE_REGISTRY, ruleRegistry);
+        ruleRegistry.put(MaskingLogMessageConverter.conversionWord, MaskingLogMessageConverter.class.getName());
+        ruleRegistry.put(LogMessageTypeConverter.conversionWord, LogMessageTypeConverter.class.getName());
+        return lc;
     }
 
     private void configureLogFile(Logger rootLogger, String logFile) {
@@ -93,13 +121,14 @@ public final class FortifyCLIDynamicInitializer {
         FileAppender<ILoggingEvent> fileAppender = new FileAppender<ILoggingEvent>();
         fileAppender.setFile(logFile);
         fileAppender.setAppend(false);
-        LayoutWrappingEncoder<ILoggingEvent> encoder = new LayoutWrappingEncoder<ILoggingEvent>();
-        encoder.setContext(loggerContext);
-        TTLLLayout layout = new TTLLLayout();
-        layout.setContext(loggerContext);
-        layout.start();
-        encoder.setLayout(layout);
-        fileAppender.setEncoder(encoder);
+        //LayoutWrappingEncoder<ILoggingEvent> encoder = new LayoutWrappingEncoder<ILoggingEvent>();
+        //encoder.setContext(loggerContext);
+        var ple = new PatternLayoutEncoder();
+        ple.setPattern("%d{HH:mm:ss.SSS} [%thread] %-5level %"+LogMessageTypeConverter.conversionWord+" %logger{36} -%kvp- %"+MaskingLogMessageConverter.conversionWord+"%n");
+        ple.setContext(loggerContext);
+        ple.start();
+        //encoder.setLayout(layout);
+        fileAppender.setEncoder(ple);
         fileAppender.setContext(loggerContext);
         fileAppender.start();
         rootLogger.addAppender(fileAppender);