+++ fcli-core/fcli-common/src/main/java/com/fortify/cli/common/action/cli/cmd/AbstractActionAsciidocCommand.java
@@ -52,7 +52,7 @@ public abstract class AbstractActionAsciidocCommand extends AbstractRunnableComm
     
     @Override @SneakyThrows
     public final Integer call() {
-        initMixins();
+        initialize();
         var contents = generateHeader();
         contents += ActionLoaderHelper
             .streamAsActions(actionSourceResolver.getActionSources(getType()), ActionValidationHandler.IGNORE)