+++ fcli-core/fcli-common/src/main/java/com/fortify/cli/common/action/cli/cmd/AbstractActionHelpCommand.java
@@ -46,7 +46,7 @@ public abstract class AbstractActionHelpCommand extends AbstractRunnableCommand 
     
     @Override
     public final Integer call() {
-        initMixins();
+        initialize();
         var action = actionResolver.loadAction(getType(), ActionValidationHandler.WARN);
         System.out.println(getActionHelp(action));
         return 0;