+++ fcli-core/fcli-sc-sast/src/main/java/com/fortify/cli/sc_sast/scan/cli/cmd/SCSastScanStartCommand.java
@@ -28,6 +28,9 @@ import com.fasterxml.jackson.databind.JsonNode;
 import com.fortify.cli.common.cli.mixin.CommonOptionMixins;
 import com.fortify.cli.common.exception.FcliBugException;
 import com.fortify.cli.common.exception.FcliSimpleException;
+import com.fortify.cli.common.log.LogMaskHelper;
+import com.fortify.cli.common.log.LogSensitivityLevel;
+import com.fortify.cli.common.log.MaskValue;
 import com.fortify.cli.common.output.cli.mixin.OutputHelperMixins;
 import com.fortify.cli.common.output.transform.IActionCommandResultSupplier;
 import com.fortify.cli.common.util.DebugHelper;
@@ -197,7 +200,9 @@ public final class SCSastScanStartCommand extends AbstractSCSastJsonNodeOutputCo
         @Getter private String appVersionNameOrId;
         @Option(names = {"--publish-as"}, required = false)
         @Getter private String fprFileName = "";
-        @Option(names = {"--publish-token"}) private String publishToken;
+        @Option(names = {"--publish-token"})
+        @MaskValue(sensitivity = LogSensitivityLevel.high, description = "SC-SAST PUBLISH TOKEN")
+        private String publishToken;
         public final boolean hasAppVersion() { return StringUtils.isNotBlank(appVersionNameOrId); }
     }