+++ fcli-other/fcli-docker/gitlab-installer-svc/linux/scripts/entrypoint.sh
@@ -1,12 +1,12 @@
 #!/bin/sh
 set -o errexit
-if [ "true" = "${CI_DEBUG_TRACE}" ]; then
+if [ "true" = "${CI_DEBUG_SERVICES}" ]; then
   set -x
 fi
 
 INSTALL_DIR=$1
 
-echo "Install fcli: $($FCLI_DIR/fcli -V)"
+echo "Install fcli: $($FCLI_DIR/fcli -V) to $INSTALL_DIR"
 cp "$FCLI_DIR/fcli" "$INSTALL_DIR"
 
 # Allow GitLab service readyness probe to succeed